back_end/requirements.txt
asgiref==3.8.1
azure-core==1.31.0
azure-storage-blob==12.23.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.0.6
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
Faker==30.8.2
google-cloud-storage==2.18.0
gunicorn[gevent]==23.0.0
idna==3.10
isodate==0.6.1
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.8.2
google-cloud-storage==2.18.0
pybtex==0.25.1
django-filter==25.1
pydantic==2.11.9
google-genai==1.38.0
werkzeug==3.0.1 
bleach==6.3.0